完成积分商城除订单支付

+++ core/src/main/java/cn/wolfcode/trip/base/domain/Goods.java
@@ -19,7 +19,7 @@ public class Goods {
 
     private String introduce;
 
-    private BigDecimal money;
+    private Integer money;
 
     private Integer sellcount;
